Moved link handleRemove so its handled in radixAutoFocus
