+++ src/components/SlateEditor/plugins/link/Link.tsx
@@ -8,9 +8,8 @@
 
 import { useCallback, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { BaseSelection, Editor, Node, Transforms } from "slate";
+import { BaseSelection, Editor, Node, Transforms, Element } from "slate";
 import { ReactEditor, RenderElementProps, useSelected } from "slate-react";
-import { css } from "@emotion/react";
 import styled from "@emotion/styled";
 import { Portal } from "@radix-ui/react-portal";
 import { ButtonV2 } from "@ndla/button";
@@ -18,6 +17,7 @@ import { colors, spacing, stackOrder } from "@ndla/core";
 import { Modal, ModalContent, ModalTrigger } from "@ndla/modal";
 import { ContentLinkElement, LinkElement } from ".";
 import EditLink from "./EditLink";
+import { TYPE_CONTENT_LINK, TYPE_LINK } from "./types";
 import config from "../../../../config";
 import { toEditGenericArticle, toLearningpathFull } from "../../../../util/routeHelpers";
 import { useArticleLanguage } from "../../ArticleLanguageProvider";
@@ -116,6 +116,14 @@ const Link = ({ attributes, editor, element, children }: Props) => {
     [focusEditor],
   );
 
+  const handleRemove = () => {
+    Transforms.unwrapNodes(editor, {
+      match: (node) => Element.isElement(node) && (node.type === TYPE_LINK || node.type === TYPE_CONTENT_LINK),
+      mode: "lowest",
+    });
+    ReactEditor.focus(editor);
+  };
+
   useEffect(() => {
     const setStateFromNode = async () => {
       let href;
@@ -172,12 +180,21 @@ const Link = ({ attributes, editor, element, children }: Props) => {
       <ModalContent
         onEscapeKeyDown={(e) => e.stopPropagation()}
         onCloseAutoFocus={(e) => {
+          if (!model?.href) {
+            handleRemove();
+          }
           e.preventDefault();
           focusEditor();
         }}
       >
         {model && (
-          <EditLink editor={editor} element={element} model={model} closeEditMode={() => toggleEditMode(false)} />
+          <EditLink
+            editor={editor}
+            element={element}
+            model={model}
+            closeEditMode={() => toggleEditMode(false)}
+            handleRemove={handleRemove}
+          />
         )}
       </ModalContent>
     </Modal>
